HHH-16104 Clarify that tests in package org.hibernate.orm.test.cdi.general.nonregistrymanaged are about Hibernate Search
By renaming that package to
org.hibernate.orm.test.cdi.general.hibernatesearch, renaming the test
classes to include "HibernateSearch" in their name, and adapting
comments.
