+++ hibernate-core/src/test/java/org/hibernate/orm/test/cdi/general/hibernatesearch/TheAlternativeNamedApplicationScopedBeanImpl.java
@@ -4,37 +4,40 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
  */
-package org.hibernate.orm.test.cdi.general.nonregistrymanaged;
+package org.hibernate.orm.test.cdi.general.hibernatesearch;
 
 import jakarta.annotation.PostConstruct;
 import jakarta.annotation.PreDestroy;
 import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Named;
 
 /**
  * @author Yoann Rodiere
  */
 @ApplicationScoped
-public class TheApplicationScopedBean {
+@Named(TheAlternativeNamedApplicationScopedBeanImpl.NAME)
+public class TheAlternativeNamedApplicationScopedBeanImpl implements TheNamedApplicationScopedBean {
+	public static final String NAME = "TheAlternativeNamedApplicationScopedBeanImpl_name";
 
 	@jakarta.inject.Inject
 	private TheNestedDependentBean nestedDependentBean;
 
-	public TheApplicationScopedBean() {
-		Monitor.theApplicationScopedBean().instantiated();
+	public TheAlternativeNamedApplicationScopedBeanImpl() {
+		Monitor.theAlternativeNamedApplicationScopedBean().instantiated();
 	}
 
+	@Override
 	public void ensureInitialized() {
 		nestedDependentBean.ensureInitialized();
 	}
 
 	@PostConstruct
 	public void postConstruct() {
-		Monitor.theApplicationScopedBean().postConstructCalled();
+		Monitor.theAlternativeNamedApplicationScopedBean().postConstructCalled();
 	}
 
 	@PreDestroy
 	public void preDestroy() {
-		Monitor.theApplicationScopedBean().preDestroyCalled();
+		Monitor.theAlternativeNamedApplicationScopedBean().preDestroyCalled();
 	}
-
 }
