+++ hibernate-core/src/test/java/org/hibernate/orm/test/cdi/general/hibernatesearch/TheApplicationScopedBean.java
@@ -4,40 +4,37 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
  */
-package org.hibernate.orm.test.cdi.general.nonregistrymanaged;
+package org.hibernate.orm.test.cdi.general.hibernatesearch;
 
 import jakarta.annotation.PostConstruct;
 import jakarta.annotation.PreDestroy;
-import jakarta.enterprise.context.Dependent;
-import jakarta.inject.Named;
+import jakarta.enterprise.context.ApplicationScoped;
 
 /**
  * @author Yoann Rodiere
  */
-@Dependent
-@Named(TheMainNamedDependentBeanImpl.NAME)
-public class TheMainNamedDependentBeanImpl implements TheNamedDependentBean {
-	public static final String NAME = "TheMainNamedDependentBeanImpl_name";
+@ApplicationScoped
+public class TheApplicationScopedBean {
 
 	@jakarta.inject.Inject
 	private TheNestedDependentBean nestedDependentBean;
 
-	public TheMainNamedDependentBeanImpl() {
-		Monitor.theMainNamedDependentBean().instantiated();
+	public TheApplicationScopedBean() {
+		Monitor.theApplicationScopedBean().instantiated();
 	}
 
-	@Override
 	public void ensureInitialized() {
 		nestedDependentBean.ensureInitialized();
 	}
 
 	@PostConstruct
 	public void postConstruct() {
-		Monitor.theMainNamedDependentBean().postConstructCalled();
+		Monitor.theApplicationScopedBean().postConstructCalled();
 	}
 
 	@PreDestroy
 	public void preDestroy() {
-		Monitor.theMainNamedDependentBean().preDestroyCalled();
+		Monitor.theApplicationScopedBean().preDestroyCalled();
 	}
+
 }
