+++ hibernate-core/src/test/java/org/hibernate/orm/test/cdi/general/hibernatesearch/TheMainNamedApplicationScopedBeanImpl.java
@@ -4,37 +4,40 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
  */
-package org.hibernate.orm.test.cdi.general.nonregistrymanaged;
+package org.hibernate.orm.test.cdi.general.hibernatesearch;
 
 import jakarta.annotation.PostConstruct;
 import jakarta.annotation.PreDestroy;
 import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Named;
 
 /**
  * @author Yoann Rodiere
  */
 @ApplicationScoped
-public class TheSharedApplicationScopedBean {
+@Named(TheMainNamedApplicationScopedBeanImpl.NAME)
+public class TheMainNamedApplicationScopedBeanImpl implements TheNamedApplicationScopedBean {
+	public static final String NAME = "TheMainNamedApplicationScopedBeanImpl_name";
 
 	@jakarta.inject.Inject
 	private TheNestedDependentBean nestedDependentBean;
 
-	public TheSharedApplicationScopedBean() {
-		Monitor.theSharedApplicationScopedBean().instantiated();
+	public TheMainNamedApplicationScopedBeanImpl() {
+		Monitor.theMainNamedApplicationScopedBean().instantiated();
 	}
 
+	@Override
 	public void ensureInitialized() {
 		nestedDependentBean.ensureInitialized();
 	}
 
 	@PostConstruct
 	public void postConstruct() {
-		Monitor.theSharedApplicationScopedBean().postConstructCalled();
+		Monitor.theMainNamedApplicationScopedBean().postConstructCalled();
 	}
 
 	@PreDestroy
 	public void preDestroy() {
-		Monitor.theSharedApplicationScopedBean().preDestroyCalled();
+		Monitor.theMainNamedApplicationScopedBean().preDestroyCalled();
 	}
-
 }
