+++ hibernate-core/src/test/java/org/hibernate/orm/test/cdi/general/hibernatesearch/TheMainNamedDependentBeanImpl.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
  */
-package org.hibernate.orm.test.cdi.general.nonregistrymanaged;
+package org.hibernate.orm.test.cdi.general.hibernatesearch;
 
 import jakarta.annotation.PostConstruct;
 import jakarta.annotation.PreDestroy;
@@ -15,15 +15,15 @@ import jakarta.inject.Named;
  * @author Yoann Rodiere
  */
 @Dependent
-@Named(TheAlternativeNamedDependentBeanImpl.NAME)
-public class TheAlternativeNamedDependentBeanImpl implements TheNamedDependentBean {
-	public static final String NAME = "TheAlternativeNamedDependentBeanImpl_name";
+@Named(TheMainNamedDependentBeanImpl.NAME)
+public class TheMainNamedDependentBeanImpl implements TheNamedDependentBean {
+	public static final String NAME = "TheMainNamedDependentBeanImpl_name";
 
 	@jakarta.inject.Inject
 	private TheNestedDependentBean nestedDependentBean;
 
-	public TheAlternativeNamedDependentBeanImpl() {
-		Monitor.theAlternativeNamedDependentBean().instantiated();
+	public TheMainNamedDependentBeanImpl() {
+		Monitor.theMainNamedDependentBean().instantiated();
 	}
 
 	@Override
@@ -33,11 +33,11 @@ public class TheAlternativeNamedDependentBeanImpl implements TheNamedDependentBe
 
 	@PostConstruct
 	public void postConstruct() {
-		Monitor.theAlternativeNamedDependentBean().postConstructCalled();
+		Monitor.theMainNamedDependentBean().postConstructCalled();
 	}
 
 	@PreDestroy
 	public void preDestroy() {
-		Monitor.theAlternativeNamedDependentBean().preDestroyCalled();
+		Monitor.theMainNamedDependentBean().preDestroyCalled();
 	}
 }
