+++ hibernate-core/src/test/java/org/hibernate/orm/test/cdi/general/hibernatesearch/TheSharedApplicationScopedBean.java
@@ -4,40 +4,37 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
  */
-package org.hibernate.orm.test.cdi.general.nonregistrymanaged;
+package org.hibernate.orm.test.cdi.general.hibernatesearch;
 
 import jakarta.annotation.PostConstruct;
 import jakarta.annotation.PreDestroy;
 import jakarta.enterprise.context.ApplicationScoped;
-import jakarta.inject.Named;
 
 /**
  * @author Yoann Rodiere
  */
 @ApplicationScoped
-@Named(TheAlternativeNamedApplicationScopedBeanImpl.NAME)
-public class TheAlternativeNamedApplicationScopedBeanImpl implements TheNamedApplicationScopedBean {
-	public static final String NAME = "TheAlternativeNamedApplicationScopedBeanImpl_name";
+public class TheSharedApplicationScopedBean {
 
 	@jakarta.inject.Inject
 	private TheNestedDependentBean nestedDependentBean;
 
-	public TheAlternativeNamedApplicationScopedBeanImpl() {
-		Monitor.theAlternativeNamedApplicationScopedBean().instantiated();
+	public TheSharedApplicationScopedBean() {
+		Monitor.theSharedApplicationScopedBean().instantiated();
 	}
 
-	@Override
 	public void ensureInitialized() {
 		nestedDependentBean.ensureInitialized();
 	}
 
 	@PostConstruct
 	public void postConstruct() {
-		Monitor.theAlternativeNamedApplicationScopedBean().postConstructCalled();
+		Monitor.theSharedApplicationScopedBean().postConstructCalled();
 	}
 
 	@PreDestroy
 	public void preDestroy() {
-		Monitor.theAlternativeNamedApplicationScopedBean().preDestroyCalled();
+		Monitor.theSharedApplicationScopedBean().preDestroyCalled();
 	}
+
 }
